migrate route registration
