+++ packages/data-access-plugin/src/manager/index.ts
@@ -2,10 +2,9 @@ import { Express } from 'express';
 import { Client } from 'elasticsearch';
 import * as apollo from 'apollo-server-express';
 import { Context } from '@terascope/job-components';
-import { Logger, toBoolean, get, isProd } from '@terascope/utils';
-import { ACLManager, User } from '@terascope/data-access';
+import { Logger, get, isProd } from '@terascope/utils';
+import { ACLManager } from '@terascope/data-access';
 import { TeraserverConfig, PluginConfig } from '../interfaces';
-import { makeSearchFn } from '../search/utils';
 import { makeErrorHandler, getESClient } from '../utils';
 import typeDefs from './types';
 import resolvers from './resolvers';
@@ -155,43 +154,6 @@ export default class ManagerPlugin {
             },
             app: this.app,
             path: managerUri,
-        });
-
-        // this must happen at the end
-        this.app.use('/api/v2/:endpoint', (req, res, next) => {
-            const manager: ACLManager = get(req, 'aclManager');
-            const user: User = utils.getLoggedInUser(req)!;
-
-            const { endpoint } = req.params;
-            const logger = this.context.apis.foundation.makeLogger({
-                module: `search_plugin:${endpoint}`,
-                user_id: get(user, 'id')
-            });
-
-            const spaceErrorHandler = makeErrorHandler('Error accessing search endpoint', logger, true);
-
-            spaceErrorHandler(req, res, async () => {
-                const accessConfig = await manager.getViewForSpace({
-                    space: endpoint
-                }, user);
-
-                req.query.pretty = toBoolean(req.query.pretty);
-
-                const connection = get(accessConfig, 'search_config.connection', 'default');
-                const client = getESClient(this.context, connection);
-
-                const search = makeSearchFn(client, accessConfig, logger);
-
-                // @ts-ignore
-                req.space = {
-                    searchErrorHandler: makeErrorHandler('Error during query execution', logger, true),
-                    accessConfig,
-                    search,
-                    logger,
-                };
-
-                next();
-            });
         });
     }
 }
