+++ packages/data-access-plugin/src/search/index.ts
@@ -1,10 +1,13 @@
 import { Express } from 'express';
 import { Client } from 'elasticsearch';
-import { Logger } from '@terascope/utils';
-import { DataAccessConfig } from '@terascope/data-access';
+import { Context } from '@terascope/job-components';
+import { Logger, toBoolean, get } from '@terascope/utils';
+import { DataAccessConfig, User, ACLManager } from '@terascope/data-access';
 import { TeraserverConfig, PluginConfig } from '../interfaces';
 import { SearchFn } from './interfaces';
-import { ErrorHandlerFn } from '../utils';
+import { ErrorHandlerFn, getESClient, makeErrorHandler } from '../utils';
+import { makeSearchFn } from './utils';
+import * as managerUtils from '../manager/utils';
 
 /**
  * @todo add support for the search counter
@@ -14,12 +17,14 @@ export default class SearchPlugin {
     readonly logger: Logger;
     readonly app: Express;
     readonly client: Client;
+    readonly context: Context;
 
     constructor(pluginConfig: PluginConfig) {
         this.client = pluginConfig.elasticsearch;
         this.config = pluginConfig.server_config;
         this.logger = pluginConfig.logger;
         this.app = pluginConfig.app;
+        this.context = pluginConfig.context;
     }
 
     async initialize() {}
@@ -50,6 +55,45 @@ export default class SearchPlugin {
                 } else {
                     res.json(result);
                 }
+            });
+        });
+    }
+
+    postRegistration() {
+        // this must happen at the end
+        this.app.use('/api/v2/:endpoint', (req, res, next) => {
+            const manager: ACLManager = get(req, 'aclManager');
+            const user: User = managerUtils.getLoggedInUser(req)!;
+
+            const { endpoint } = req.params;
+            const logger = this.context.apis.foundation.makeLogger({
+                module: `search_plugin:${endpoint}`,
+                user_id: get(user, 'id')
+            });
+
+            const spaceErrorHandler = makeErrorHandler('Error accessing search endpoint', logger, true);
+
+            spaceErrorHandler(req, res, async () => {
+                const accessConfig = await manager.getViewForSpace({
+                    space: endpoint
+                }, user);
+
+                req.query.pretty = toBoolean(req.query.pretty);
+
+                const connection = get(accessConfig, 'search_config.connection', 'default');
+                const client = getESClient(this.context, connection);
+
+                const search = makeSearchFn(client, accessConfig, logger);
+
+                // @ts-ignore
+                req.space = {
+                    searchErrorHandler: makeErrorHandler('Error during query execution', logger, true),
+                    accessConfig,
+                    search,
+                    logger,
+                };
+
+                next();
             });
         });
     }
